Locking enhancements
- spin_try_lock_irq_save
- cleanup to avoid warnings on recent compilers

+++ include/nautilus/cpu_state.h
@@ -48,7 +48,7 @@ static inline void preempt_disable()
     void *base = __cpu_state_get_cpu();
     if (base) {
 	// per-cpu functional
-	__sync_fetch_and_add((uint16_t *)(base+10),1);
+	__sync_fetch_and_add((uint16_t *)((uint64_t)base+10),1);
     } else {
 	// per-cpu is not running, so we are not going to get preempted anyway
     }
@@ -59,7 +59,7 @@ static inline void preempt_enable()
     void *base = __cpu_state_get_cpu();
     if (base) {
 	// per-cpu functional
-	__sync_fetch_and_sub((uint16_t *)(base+10),1);
+	__sync_fetch_and_sub((uint16_t *)((uint64_t)base+10),1);
     } else {
 	// per-cpu is not running, so we are not going to get preempted anyway
     }
@@ -73,7 +73,7 @@ static inline void preempt_reset()
     void *base = __cpu_state_get_cpu();
     if (base) {
 	// per-cpu functional
-	__sync_fetch_and_and((uint16_t *)(base+10),0);
+	__sync_fetch_and_and((uint16_t *)((uint64_t)base+10),0);
     } else {
 	// per-cpu is not running, so we are not going to get preempted anyway
     }
@@ -84,7 +84,7 @@ static inline int preempt_is_disabled()
     void *base = __cpu_state_get_cpu();
     if (base) {
 	// per-cpu functional
-	return __sync_fetch_and_add((uint16_t *)(base+10),0);
+	return __sync_fetch_and_add((uint16_t *)((uint64_t)base+10),0);
     } else {
 	// per-cpu is not running, so we are not going to get preempted anyway
 	return 1;
@@ -95,7 +95,7 @@ static inline uint16_t interrupt_nesting_level()
 {
     void *base = __cpu_state_get_cpu();
     if (base) {
-	return __sync_fetch_and_add((uint16_t *)(base+8),0);
+	return __sync_fetch_and_add((uint16_t *)((uint64_t)base+8),0);
     } else {
 	return 0; // no interrupt should be on if we don't have percpu
     }
